Finally added simple max_memory_used detection
Maybe it should be tested on speed (as long as we are reading values
from /proc, it could be not very speed) and on correctness (it usually
shows a few hundred bytes more than valgrind does). Those isues will be
soon reported on github. After their close, that code will be mostly
moved to utils and pushed into master.